refactor: allocate exception handler in utils module
